completed basic api: (see note)
Basic api for app is complete. Of course it needs more work but the aim was to create somthing that i could work with in the meantime whilst i tackle angular on teh front end. Will add to api as i need

+++ src/controllers/postComments.ts
@@ -1,6 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import CommentsModel from "../models/Comments";
 import Logging from "../library/Logging";
+import { iComments } from "../customTypings/interfaces/schema and model/iComments";
 
 export const getPostComments = async (
   req: Request,
@@ -56,17 +57,30 @@ export const deletComment = async (
   res: Response,
   next: NextFunction
 ) => {
-  const commentId = req.params.commentId;
+  const commentId = req.query.commentId;
+  const commentAuthor = req.query.comentator;
+  const postAuthor = req.query.postAuthorId;
 
   try {
-    const result = await CommentsModel.deleteOne({ _id: commentId });
-    if (result.deletedCount > 0) {
-      res.status(204).end();
+    const comment = await CommentsModel.findById(commentId);
+    if (!comment) {
+      res.status(404).json({ message: "Comment not found" });
+    } else if (
+      (comment as iComments).comentatorId.toString() === commentAuthor ||
+      postAuthor === (req.user as any).id
+    ) {
+      const result = await CommentsModel.deleteOne({ _id: commentId });
+      if (result.deletedCount > 0) {
+        res.status(204).end();
+      } else throw new Error("Comment failed to delete");
     } else {
-      throw new Error("Failed to delete comment");
+      return res.status(403).json({ message: "Unathorized action" });
     }
   } catch (error) {
     Logging.error(error);
-    res.status(500).json({ message: "Internal server error", error: error });
+    res.status(500).json({
+      message: "Internal server error",
+      error: error instanceof Error ? error.message : error,
+    });
   }
 };
